requirements.txt
numpy
onnxruntime-gpu
Pillow
pyinstaller